refactor(auth): extract reusable fixed-window ip limiter
