+++ backend/src/test/java/com/univoyage/auth/security/FixedWindowIpRateLimiterTest.java
@@ -1,7 +1,5 @@
 package com.univoyage.auth.security;
 
-import com.univoyage.auth.config.LoginSecurityProperties;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -15,22 +13,13 @@ import java.util.concurrent.atomic.AtomicLong;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-class LoginIpRateLimiterTest {
-
-    private LoginSecurityProperties properties;
-
-    @BeforeEach
-    void setUp() {
-        properties = new LoginSecurityProperties();
-        properties.setIpMaxAttempts(3);
-        properties.setIpWindow(Duration.ofMinutes(1));
-    }
+class FixedWindowIpRateLimiterTest {
 
     @Test
     @DisplayName("Returns -1 while under max attempts for the same IP in the current window")
     void allowsUpToMaxAttempts() {
         Clock clock = Clock.fixed(Instant.parse("2026-01-01T00:00:00Z"), ZoneOffset.UTC);
-        LoginIpRateLimiter limiter = new LoginIpRateLimiter(properties, clock);
+        FixedWindowIpRateLimiter limiter = new FixedWindowIpRateLimiter(clock, 3, Duration.ofMinutes(1));
 
         assertEquals(-1, limiter.tryConsumeOrRetryAfterSeconds("192.168.1.1"));
         assertEquals(-1, limiter.tryConsumeOrRetryAfterSeconds("192.168.1.1"));
@@ -41,7 +30,7 @@ class LoginIpRateLimiterTest {
     @DisplayName("After max attempts, returns positive retry-after seconds until window end")
     void blocksWithRetryAfterWhenOverLimit() {
         Clock clock = Clock.fixed(Instant.parse("2026-01-01T00:00:00Z"), ZoneOffset.UTC);
-        LoginIpRateLimiter limiter = new LoginIpRateLimiter(properties, clock);
+        FixedWindowIpRateLimiter limiter = new FixedWindowIpRateLimiter(clock, 3, Duration.ofMinutes(1));
 
         assertEquals(-1, limiter.tryConsumeOrRetryAfterSeconds("10.0.0.1"));
         assertEquals(-1, limiter.tryConsumeOrRetryAfterSeconds("10.0.0.1"));
@@ -55,7 +44,7 @@ class LoginIpRateLimiterTest {
     @DisplayName("Different IPs have independent counters")
     void separateCountersPerIp() {
         Clock clock = Clock.fixed(Instant.parse("2026-06-01T12:00:00Z"), ZoneOffset.UTC);
-        LoginIpRateLimiter limiter = new LoginIpRateLimiter(properties, clock);
+        FixedWindowIpRateLimiter limiter = new FixedWindowIpRateLimiter(clock, 3, Duration.ofMinutes(1));
 
         for (int i = 0; i < 3; i++) {
             assertEquals(-1, limiter.tryConsumeOrRetryAfterSeconds("ip-a"));
@@ -69,7 +58,7 @@ class LoginIpRateLimiterTest {
     @DisplayName("Blank IP is normalized to a single bucket key")
     void blankIpUsesUnknownBucket() {
         Clock clock = Clock.fixed(Instant.parse("2026-06-01T12:00:00Z"), ZoneOffset.UTC);
-        LoginIpRateLimiter limiter = new LoginIpRateLimiter(properties, clock);
+        FixedWindowIpRateLimiter limiter = new FixedWindowIpRateLimiter(clock, 3, Duration.ofMinutes(1));
 
         assertEquals(-1, limiter.tryConsumeOrRetryAfterSeconds(""));
         assertEquals(-1, limiter.tryConsumeOrRetryAfterSeconds("   "));
@@ -99,8 +88,7 @@ class LoginIpRateLimiterTest {
             }
         };
 
-        properties.setIpWindow(Duration.ofMillis(10_000));
-        LoginIpRateLimiter limiter = new LoginIpRateLimiter(properties, clock);
+        FixedWindowIpRateLimiter limiter = new FixedWindowIpRateLimiter(clock, 3, Duration.ofMillis(10_000));
 
         assertEquals(-1, limiter.tryConsumeOrRetryAfterSeconds("moving"));
         assertEquals(-1, limiter.tryConsumeOrRetryAfterSeconds("moving"));
